requirements2.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.16.0
chardet==4.0.0
cryptography==41.0.5
defusedxml==0.7.1
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.3
fonttools==4.45.1
fpdf2==2.7.6
idna==2.10
lxml==4.9.3
mailjet-rest==1.3.4
mysql-client==0.0.1
mysqlclient==2.0.3
Pillow==10.1.0
pycparser==2.21
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
suds==1.1.2
urllib3==1.26.4
xmltodict==0.13.0
